- extracted RuntimeHelpers.GetSubArray to spearate files;
