+++ src/Stl.Testing/Internal/TestOutputWriter.cs
@@ -48,22 +48,3 @@ namespace Stl.Testing.Internal
     }
 
 }
-
-
-#if NETSTANDARD2_0
-namespace System.Runtime.CompilerServices
-{
-    public static class RuntimeHelpers
-    {
-        public static T[] GetSubArray<T>(T[] array, System.Range range)
-        {
-            (int offset, int length) = range.GetOffsetAndLength(array.Length);
-            var arr = new T[length];
-            for (int i = 0; i < length; i++) {
-                arr[i] = array[offset + i];
-            }
-            return arr;
-        }
-    }
-}
-#endif
